fix: Add fall-through for missing Docker-Content-Digest (release-3.9)
PR sylabs/singularity#727 was written while reviewing the Docker registry API v2 spec,
in which it is stated a manifest HEAD request does include
`Docker-Content-Digest`.

I had not considered potential differences in the
oci-distribution-spec. This makes it clear that
`Docker-Content-Digest` is a SHOULD header, and lists it as a field
clients should not depend on.

The `containers/image` `docker.GetDigest` function doesn't handle the
lack of a `Docker-Content-Digest` header, so we should fall-through to
the old `GetManifest` approach on error.

Fixes sylabs/singularity#733

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>