fix: Check image can be read before opening
If an image cannot be read, attempts to open it with the various image
format handling code will fail. Since we are using a fall-through
approach to find the correct image format, these failures did not give
a sensible error message.

Add an explicit check that the image is readable.

Fixes sylabs/singularity#786

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>