feat: exec / run args support for --oci mode
When using `run` or `exec` with the `--oci` runtime mode, accept
arguments on the command line.

For `run`, the arguments override any CMD specified by the image.

For `exec`, the arguments replace ENTRYPOINT/CMD entirely, bypassing
the process configuration in the image config.

This mirrors the behavior of Singularity images today, via the exec
and run runscripts - but is implemented in the OCI bundle config,
rather than a script in the container.

Closes sylabs/singularity#1024

Closes sylabs/singularity#1092

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>