chore: refactor runtime spec handling out of bundle
Bring all handling of the runtime spec out of the native oci bundle
package, up into the oci launcher.

We obtain the bundle's image spec after it is downloaded / extracted.
The launcher then computes the correct Process config and updates the
bundle with it.

This is required so that we can handle an image USER in the launcher.

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>