fix: fall back to cgroupfs for OCI non-root, non cgroups v2
For the `OCI` command group, we need cgroups v2 unified mode in order
to call `runc` with `--systemd-cgroup` as a non-root user.

Fall back to cgroupfs operation if this constraint is not satisified.

This fixes a failure when running a container using `runc`. `crun`
was falling back to cgroupfs itself, even when `--systemd-cgroup` was
specified.

Fixes sylabs/singularity#1408

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>