fix: oci: explicitly request userns for inner reverse idmap
When running in --oci mode as a normal user, Singularity sets up an
outer userns that provides an unprivileged root. When --fakeroot is
not also requested, runc / crun then map back to the user's own
uid/gid via an inner reverse mapping specified in config.json.

This mapping was not being applied correctly, as a userns was not
requested in the config.json for all situations where a reverse
mapping was requested.

Ensure that if we specify a mapping in config.json then we request a
userns so that it is applied.

Cover this behavior in an e2e-test by writing to a host dir from the
container, and checking ownership of the file on the host.

Fixes sylabs/singularity#1517

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>