fix: oci: Don't create cgroup for crun on v1 / cgroupfs
If we are running under cgroups v1 or with the cgroupfs
manager (i.e. not systemd as cgroup manager), do not attempt to enter
a cgroup at startup with crun. We cannot create a cgroup unprivileged
in this situation.

Under cgroups v1, crun will not perform the cgroups manipulation that
leads to the issue we worked around in #1539. Any other issue with the
cgroup that we are in at launch cannot be rectified, either.

Fixes sylabs/singularity#1569

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>