fix: Remove files with restrictive perms from --oci temp rootfs
When we run a container in `--oci` mode, we have a temporary rootfs
that must be removed when the container exits.

A container can contain files / dirs with restrictive permissions that
prevent direct removal.

Use fs.ForceRemoveAll, which will chmod & remove anything that
os.RemoveAll is unable to remove.

Fixes sylabs/singularity#1586

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>