--no-setgroups option for --fakeroot
When we run in `--fakeroot` mode, the `starter` will call `setgroups`
and reduce the group membership of the container process to `gid=0`
only, inside the fakeroot id-mapped user namespace.

This removes supplementary group membership, and therefore prevents
access to files on the host that are restricted by group ownership
where the group is a supplementary group of the current user.

If, for example, `$HOME` is accessed via a supplementary group then it
cannot be mounted inside the id-mapped user namespace created at the
start of a `singularity build --fakeroot`. The build will then fail.

This PR adds a new flag `--no-setgroups` to skip the `setgroups` call
in `--fakeroot` mode. This results in the container process *not*
having its group membership changed. Host files accessible via host
supplementary groups will remain accessible.

This does not make the supplementary groups fully available in the
container. Inside the id-mapped user namespace they will be mapped to
`nobody`. The output of `id` may be confusing, and we still cannot
e.g. `chgrp` a file to one of the user's supplementary groups.

On host...

   $ groups
   dtrudg-sylabs wheel mock docker libvirt

In container with `--fakeroot --no-setgroups`...

   Singularity> groups
   root nobody nobody nobody nobody

Fixes sylabs/singularity#1748

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>