oci: Support --no-mount, to extent possible
Allow the `--no-mount` flag to be specified in `--oci` mode. This
allows disabling the following mounts:

* proc
* sys
* devpts
* tmp
* home

Note that `dev` cannot be supported in `--oci` mode, as an OCI
runtime *requires* that certain devices are present, and will include
them in a `/dev` tmpfs.

We currently run similar to native mode `--compat`, so we don't mount
the current working directory. Therefore, `--no-mount cwd` is not
supported.

Similarly, `--compat` infers that `bind path` entries from
`singularity.conf` are ignored. We may handle them in some way, in future.

Fixes sylabs/singularity#1781

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>