fix: oci: correct $HOME -> default cwd / passwd handling
In the refactoring performed in a10b379, the setting of the default
CWD, and creation of `/etc/passwd`, was not correctly adapted.

We need to use our parsed / processed `l.homeDest` value for the
container homeDir in all cases. With a nested root-mapped user
namespace re-exec, the CLI passed value is not correct unless it is a
custom value.

Fixes sylabs/singularity#1791

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>