oci: Support --add-caps and --drop-caps
Add support for --add-caps and --drop-caps to modify the capabilities
of the container process.

When the container user is root, these modify the
permitted/effective/bounding sets.

When the container user is non-root, the bounding set is modified, and
any explicit add-caps are added to the
permitted/effective/inheritable/ambient sets.

Fixes sylabs/singularity#1468

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>