oci: create runtime bundle within buildcfg.SESSIONDIR tmpfs
In Singularity's native mode, we assemble the runtime representation
of the container's rootfs under a tmpfs mounted on
buildcfg.SESSIONDIR. The pristine rootfs, whether from a SIF file or
an OCI source image extracted to a directory, is mounted into the
tmpfs before other overlays, customised files, host mounts etc. are
layered over it.

The location of buildcfg.SESSIONDIR is set when SingularityCE is
compiled. It is documented that it needs to be on a local
filesystem. Using a tmpfs mount over a local fs SESSIONDIR avoids
potential issues with mounts over network file systems etc. The OCI
mode is currently vulnerable to issues as it creates the runtime
bundle on TMPDIR.

Also, prior to this PR, our code assumes that an OCI container rootfs
sits directly in the runtime bundle, and can be modified directly by
the launcher code. This does not allow e.g. mounting the rootfs out of
a read only loopback image.

In preparation for enabling oci-sif mounts, this PR modifies the existing
ocibundle and launcher code to adopt the same SESSIONDIR tmpfs
pattern.

* The oci launcher mounts a tmpfs on buildcfg.SESSIONDIR.
* The runtime bundle is created inside the tmpfs.
* The ocibundle.Create() code extracts the pristine image rootfs
  outside of the SESSIONDIR bundle, and then bind mounts it in read-only.
* /etc/passwd,group,resolv.conf are now updated by creating new files
  inside the SESSIONDIR bundle, which are bind mounted over the
  versions in the pristine rootfs.

Fixes: sylabs/singularity#1855
Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>