oci: Handle oci-sif image refs without oci-sif: prefix
Allow `run/shell/exec --oci` to handle a bare image path.

We have to check whether the image is an oci-sif in order to give a
friendly error message.

At present this is all implemented within the oci launcher, reached
only when `--oci` is specified on the CLI.

At a later date we will allow `--oci` to be implied by attempting to
execute an oci-sif image, so the logic will move. However, this is an
appropriate place until that behaviour is well specified. Much thought
is needed around sysadmin configurable defaults / overrides for
runtime engine selection.

Closes sylabs/singularity#1861

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>