chore: Move library push impl to internal/pkg/client/library
Implementation of pull from / push to for network sources is kept
under internal/pkg/client/<source>.

Prior to this PR, the implementation of push to a library://
destination was an exception, with the code at
internal/app/singularity

Move and refactor that code, so it fits into the client/library
package.

This is necessary to keep the upcoming OCI registry and library
OCI-SIF push implementation tidy.

Signed-off-by: Edita Kizinevic <edita.kizinevic@cern.ch>